退会条件付け実装

+++ src/pages/users/[id]/edit/leave.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useMemo } from 'react'
 import { useState } from 'react'
 import { Form, FormGroup, Button, Label, Input, Row } from 'reactstrap'
 import errorMsg from '@/src/lib/errorMsg'
@@ -6,35 +6,54 @@ import { useAlert } from "react-alert"
 import Link from 'next/link'
 import { fuego } from '@nandorojo/swr-firestore'
 import withAuth from '@/src/lib/withAuth'
+import { GetServerSideProps, NextPage } from 'next'
+import { stripeAccounts, stripeBalance } from '@/src/lib/stripeRetrieve'
+import { Stripe } from '@/src/lib/stripe'
+import initFirebaseAdmin from '@/src/lib/initFirebaseAdmin'
+import { event } from 'app'
+import moment from 'moment'
 
-export const Leave = ({user}) => {
+export const Leave: NextPage<{user: firebase.User, balance: Stripe.Balance, canLeave: boolean}> = ({ user, balance, canLeave }) => {
   const alert = useAlert()
   const [checkBox, setCheckBox] = useState(false)
   const [pwd, setPwd] = useState('')
 
+  const available = useMemo(() => {
+    let sum = 0
+    balance.available.forEach(balance => sum += balance.amount)
+    return sum;
+  }, [balance.available]);
+
+  const pending = useMemo( () => {
+    let sum = 0;
+    balance.pending.forEach((balance) => (sum += balance.amount));
+    return sum;
+  }, [balance.pending]);
+
   const submit = async (e) => {
     e.preventDefault();
     if (!checkBox) return alert.error('チェックボックスが選択されていません。')
-    const { providerData } = user as firebase.User
+    if (available || pending) return alert.error('送金待ちの残高が残っています。')
+    if (!canLeave) return alert.error("開催イベントの終了から10日間は退会できません。");
+    const { providerData } = user as firebase.User;
     try {
-        if (providerData[0].providerId === "password") {
-          const credencial: firebase.auth.AuthCredential = fuego.auth.EmailAuthProvider.credential(
-            user.email,
-            pwd
-          );
-          await user.reauthenticateWithCredential(credencial);
-        } else {
-          await user.reauthenticateWithPopup(providerData[0]);
-        }
-        
-        await user.delete()
-        alert.success('退会処理が完了しました。')
+      if (providerData[0].providerId === "password") {
+        const credencial: firebase.auth.AuthCredential = fuego.auth.EmailAuthProvider.credential(
+          user.email,
+          pwd
+        );
+        await user.reauthenticateWithCredential(credencial);
+      } else {
+        await user.reauthenticateWithPopup(providerData[0]);
+      }
+      
+      await user.delete()
+      alert.success('退会処理が完了しました。')
     } catch (e) {
-        alert.error(errorMsg(e))
+      alert.error(errorMsg(e))
     }
   }
   // TODO: お問い合わせページ作成
-  // TODO: 退会条件の設定
   return (
     <Form style={{ marginTop: "6.5em" }} onSubmit={submit}>
       <h5>本当に退会しますか？</h5>
@@ -75,5 +94,22 @@ export const Leave = ({user}) => {
     </Form>
   );
 }
+
+export const getServerSideProps: GetServerSideProps = async ({ query }) => {
+  const { id } = query;
+  const balance = await stripeBalance(id);
+  const { firestore } = await initFirebaseAdmin()
+  const events = (await firestore.collection('events').where('createdUser', '==', id).get()).docs
+  let canLeave = true;
+  await Promise.all(events.map(event => {
+    const endDate = moment((event.data() as event).endDate.toDate());
+    if (endDate.add(10, "days").valueOf() <= moment().valueOf())
+      canLeave = false;
+  }))
+  return {
+    props: { balance, canLeave },
+    // revalidate: 1
+  };
+};
 
 export default withAuth(Leave)
